Update setupbase from jupyter-packaging
